+++ lib/FileBody.php
@@ -19,7 +19,7 @@ class FileBody implements AggregateBody {
     }
 
     /**
-     * Returns the raw HTTP message body to be sent
+     * Retrieve the sendable Artax entity body representation
      *
      * @param \Alert\Reactor $reactor
      * @return \After\Promise
@@ -68,6 +68,38 @@ class FileBody implements AggregateBody {
     }
 
     private function generateNaiveHeaders() {
+        $future = new Future;
+        $this->generateNaiveLength()->when(function($error, $result) use ($future) {
+            if ($error) {
+                $future->fail($error);
+            } else {
+                $future->succeed(['Content-Length' => $result]);
+            }
+        });
+
+        return $future->promise();
+    }
+
+    /**
+     * Retrieve the entity body's content length
+     *
+     * @param \Alert\Reactor $reactor
+     * @return \After\Promise
+     */
+    public function getLength(Reactor $reactor) {
+        return ($reactor instanceof UvReactor)
+            ? $this->generateUvLength($reactor)
+            : $this->generateNaiveLength();
+    }
+
+    private function generateUvLength($reactor) {
+        // @TODO Implement non-blocking php-uv file size retrieval.
+        // For now we'll just use the dumb blocking version.
+        // v1.0.0 cannot be a thing until this is implemented.
+        return $this->generateNaiveLength();
+    }
+
+    private function generateNaiveLength() {
         $size = @filesize($this->path);
         if ($size === false) {
             return new Failure(new \RuntimeException(
@@ -75,6 +107,6 @@ class FileBody implements AggregateBody {
             ));
         }
 
-        return new Success(['Content-Length' => $size]);
+        return new Success($size);
     }
 }